back/requirements.txt
alembic==1.9.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.89.1
filelock==3.9.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
identify==2.5.17
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
nodeenv==1.7.0
orjson==3.8.5
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.2
psycopg2-binary==2.9.5
pydantic==1.10.4
pytest==7.2.1
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.4.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
watchfiles==0.18.1
websockets==10.4
